feat: add payment through invoices separately in the component
