+++ src/components/form/FormDonation/PaymentBlock/PaymentBlock.module.scss
@@ -102,3 +102,52 @@
     }
   }
 }
+
+.invoice {
+  @include adaptiveValue('font-size', 16, 13, 0, 1200, 390);
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  padding: 1.5rem 2rem;
+
+  &__list {
+    display: flex;
+    flex-direction: column;
+    gap: toRem(12);
+  }
+
+  &__item {
+    display: flex;
+    align-items: flex-end;
+    justify-content: space-between;
+    flex-wrap: wrap;
+    gap: 1rem;
+    padding: 1rem 1rem 1rem 1.75rem;
+  }
+
+  &__info {
+    display: flex;
+    flex-direction: column;
+    gap: toRem(12);
+  }
+
+  &__title {
+    font-weight: 600;
+  }
+
+  &__account {
+    word-break:break-all;
+  }
+
+  &__btn {
+    font-weight: 600;
+    display: flex;
+    align-items: center;
+    gap: toRem(6);
+  }
+
+  &__SWIFT {
+    font-weight: 600;
+  }
+}
